new file for types and avatarList
